notebook.py to resolve circular import
